Move shuffling functions to helpers.py
