+++ beacon_chain/state/helpers.py
@@ -1,6 +1,51 @@
+from beacon_chain.utils.blake import (
+    blake,
+)
 from beacon_chain.state.config import (
     DEFAULT_CONFIG,
 )
+
+
+def get_shuffling(seed, validator_count, sample=None, config=DEFAULT_CONFIG):
+    max_validators = config['max_validators']
+    assert validator_count <= max_validators
+
+    if validator_count == 0:
+        return []
+
+    rand_max = max_validators - max_validators % validator_count
+    o = list(range(validator_count))
+    source = seed
+    i = 0
+    maxvalue = sample if sample is not None else validator_count
+    while i < maxvalue:
+        source = blake(source)
+        for pos in range(0, 30, 3):
+            m = int.from_bytes(source[pos:pos+3], 'big')
+            remaining = validator_count - i
+            if remaining == 0:
+                break
+            if validator_count < rand_max:
+                replacement_pos = (m % remaining) + i
+                o[i], o[replacement_pos] = o[replacement_pos], o[i]
+                i += 1
+    return o[:maxvalue]
+
+
+def get_attesters_and_proposer(crystallized_state,
+                               active_state,
+                               skip_count,
+                               config=DEFAULT_CONFIG):
+    attester_count = config['attester_count']
+    attestation_count = min(len(crystallized_state.active_validators), attester_count)
+
+    indices = get_shuffling(
+        active_state.randao,
+        len(crystallized_state.active_validators),
+        attestation_count + skip_count + 1,
+        config
+    )
+    return indices[:attestation_count], indices[-1]
 
 
 def get_crosslink_shards_count(active_validators_count, config=DEFAULT_CONFIG):
